refactor code, add listener file
