+++ src/modules/swaps/dto/swap.dto.ts
@@ -2,7 +2,6 @@ export interface SwapDto {
   id: number;
   swappedAt: Date;
   accountId: string;
-  swapType: string;
   inputAssetId: string;
   outputAssetId: string;
   assetInputAmount: number;
