+++ src/modules/swaps/swaps.module.ts
@@ -1,19 +1,18 @@
 import { Module } from '@nestjs/common';
 import { SwapRepository } from './swaps.repository';
-import { DbConnectionService } from './db-connection.service';
 import { SwapsController } from './swaps.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Swap } from './entity/swaps.entity';
+import { SwapListener } from './swap.listener';
 import { SwapGateway } from './swaps.gateway';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Swap])],
   controllers: [SwapsController],
-  providers: [SwapRepository, DbConnectionService, SwapGateway],
+  providers: [SwapRepository, SwapGateway, SwapListener],
 })
 export class SwapsModule {
-  constructor(swap: SwapRepository, dbConnection: DbConnectionService) {
-    swap.writeSwapToDatabase();
-    dbConnection.connect();
+  constructor(private swapListener: SwapListener) {
+    this.swapListener.writeSwapToDatabase();
   }
 }
